bin/requirement.txt
selenium
pillow
xlrd
xlwt
pyinstaller